Renamed package flat -> primitive2d
